requirements.txt
streamlit==1.31.0
streamlit-drawable-canvas==0.9.3
numpy>=1.24.0
pillow>=11.3.0
opencv-python-headless>=4.12.0
matplotlib>=3.7.1